+++ grpc_mock/service.py
@@ -1,5 +1,4 @@
 import json
-import re
 import socket
 
 import blackboxprotobuf
@@ -7,74 +6,13 @@ import h2.config
 import h2.connection
 import h2.events
 
-from dev.proto_file_parser import ProtoParser
-
-PROTO = """
-syntax = "proto3";
-package library;
-
-
-message Author {
-  string last_name = 1;
-  string first_name = 2;
-  optional string second_name = 3;
-}
-
-message BookMetadata {
-  message Publisher {
-    string name = 1;
-  }
-
-  string name = 1;
-  int32 year = 2;
-  repeated Author authors = 3;
-  optional Publisher publisher = 4;
-}
-
-message BookAddRequest {
-  message User {
-    enum Sex {
-      MALE = 0;
-      FEMALE = 1;
-    }
-    string last_name = 1;
-    string first_name = 2;
-    optional string second_name = 3;
-    optional Sex sex = 4;
-  }
-
-  string book_uuid = 1;
-  int64 user_id = 2;
-  string timestamp = 3;
-  optional BookMetadata metadata = 4;
-  optional User user = 5;
-}
-
-message BookAddReply {
-  string transaction_uuid = 1;
-}
-
-message BookRemoveRequest {
-  string book_uuid = 1;
-  int64 user_id = 2;
-  string timestamp = 3;
-}
-
-message BookRemoveReply {
-  string transaction_uuid = 1;
-}
-
-// The library service definition.
-service Books {
-  rpc BookAddEndpoint (BookAddRequest) returns (BookAddReply) {}
-  rpc BookRemoveEndpoint (BookRemoveRequest) returns (BookRemoveReply) {}
-}
-
-"""
+from grpc_mock.proto_utils import get_proto_path_from_request, get_request_typedef_from_proto_package, parse_proto_file
+from grpc_mock.repository import get_proto
 
 
 HOST = "127.0.0.1"  # Standard loopback interface address (localhost)
 PORT = 3333  # Port to listen on (non-privileged ports are > 1023)
+
 
 def http2_send_response(conn: h2.connection.H2Connection, event):
     stream_id = event.stream_id
@@ -116,28 +54,22 @@ def http2_handle(sock):
             if isinstance(event, h2.events.DataReceived):
                 prepared_events["data"] = event
 
-        if len(prepared_events) == 2:
-            parse_grpc_data(prepared_events["data"], prepare_typedef(PROTO, prepared_events["request"]))
         if data_to_send := conn.data_to_send():
             sock.sendall(data_to_send)
 
-
-def prepare_typedef(proto: str, event: h2.events.RequestReceived):
-    parser = ProtoParser(proto)
-    typedef = parser.to_typedef()
-    path: str = ""
-    for header in event.headers:
-        if header[0] == b":path":
-            path = header[1].decode()
-            break
-    package, service, method = re.match(r"^/(.+)\.(.+)/(.+)$", path).groups()
-    return typedef[package]["services"][service]["methods"][method]["request"]
+    if len(prepared_events) == 2:
+        proto_path = get_proto_path_from_request(prepared_events["request"])
+        proto_package = parse_proto_file(get_proto(proto_path))
+        parse_grpc_data(
+            prepared_events["data"],
+            get_request_typedef_from_proto_package(proto_package, proto_path),
+        )
 
 
 def parse_grpc_data(event: h2.events.DataReceived, typedef: dict):
     message, _ = blackboxprotobuf.decode_message(event.data[5:], typedef)
     print(message)
-    print(json.dumps(message, ensure_ascii=False, indent=4))
+    # print(json.dumps(message, ensure_ascii=False, indent=4))
 
 
 with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
